Move ZipCompressor into `com.github.jengelman.gradle.plugins.shadow.tasks`
